fix getSubtreeId unit tests
